requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.0.0
redis==4.5.0
celery==5.3.4
Pillow==9.5.0
python-decouple==3.8
django-filter==23.1
django-extensions==3.2.0